Move item functions to items package
